Issue #213 - Update handler package structure
