+++ doc/source/ait.core.server.handlers.ccsds_packet_handler.rst
@@ -1,0 +1,7 @@
+ait.core.server.handlers.ccsds\_packet\_handler module
+======================================================
+
+.. automodule:: ait.core.server.handlers.ccsds_packet_handler
+    :members:
+    :undoc-members:
+    :show-inheritance:
