+++ doc/source/ait.core.server.handlers.packet_handler.rst
@@ -1,0 +1,7 @@
+ait.core.server.handlers.packet\_handler module
+===============================================
+
+.. automodule:: ait.core.server.handlers.packet_handler
+    :members:
+    :undoc-members:
+    :show-inheritance:
