+++ doc/source/ait.core.server.handlers.rst
@@ -1,0 +1,18 @@
+ait.core.server.handlers package
+================================
+
+Submodules
+----------
+
+.. toctree::
+
+   ait.core.server.handlers.ccsds_packet_handler
+   ait.core.server.handlers.packet_handler
+
+Module contents
+---------------
+
+.. automodule:: ait.core.server.handlers
+    :members:
+    :undoc-members:
+    :show-inheritance:
